+++ src/lib/createables/prayer.ts
@@ -1,0 +1,54 @@
+import { Createable } from '.';
+import itemID from '../util/itemID';
+
+const Prayer: Createable[] = [
+	{
+		name: 'Blessed spirit shield',
+		inputItems: {
+			[itemID('Spirit shield')]: 1,
+			[itemID('Holy elixir')]: 1
+		},
+		outputItems: {
+			[itemID('Blessed spirit shield')]: 1
+		},
+		prayerLevel: 85
+	},
+	{
+		name: 'Spectral spirit shield',
+		inputItems: {
+			[itemID('Blessed spirit shield')]: 1,
+			[itemID('Spectral sigil')]: 1
+		},
+		outputItems: {
+			[itemID('Spectral spirit shield')]: 1
+		},
+		prayerLevel: 90,
+		smithingLevel: 85
+	},
+	{
+		name: 'Arcane spirit shield',
+		inputItems: {
+			[itemID('Blessed spirit shield')]: 1,
+			[itemID('Arcane sigil')]: 1
+		},
+		outputItems: {
+			[itemID('Arcane spirit shield')]: 1
+		},
+		prayerLevel: 90,
+		smithingLevel: 85
+	},
+	{
+		name: 'Elysian spirit shield',
+		inputItems: {
+			[itemID('Blessed spirit shield')]: 1,
+			[itemID('Elysian sigil')]: 1
+		},
+		outputItems: {
+			[itemID('Elysian spirit shield')]: 1
+		},
+		prayerLevel: 90,
+		smithingLevel: 85
+	}
+];
+
+export default Prayer;
