+++ src/lib/createables/skillingSets.ts
@@ -1,0 +1,143 @@
+import { Createable } from '.';
+import itemID from '../util/itemID';
+
+const SkillingSets: Createable[] = [
+	{
+		name: 'Graceful',
+		inputItems: {
+			[itemID('Mark of grace')]: 260
+		},
+		outputItems: {
+			[itemID('Graceful hood')]: 1,
+			[itemID('Graceful top')]: 1,
+			[itemID('Graceful legs')]: 1,
+			[itemID('Graceful gloves')]: 1,
+			[itemID('Graceful boots')]: 1,
+			[itemID('Graceful cape')]: 1
+		}
+	},
+	{
+		name: 'Graceful hood',
+		inputItems: {
+			[itemID('Mark of grace')]: 35
+		},
+		outputItems: {
+			[itemID('Graceful hood')]: 1
+		}
+	},
+	{
+		name: 'Graceful top',
+		inputItems: {
+			[itemID('Mark of grace')]: 55
+		},
+		outputItems: {
+			[itemID('Graceful top')]: 1
+		}
+	},
+	{
+		name: 'Graceful legs',
+		inputItems: {
+			[itemID('Mark of grace')]: 60
+		},
+		outputItems: {
+			[itemID('Graceful legs')]: 1
+		}
+	},
+	{
+		name: 'Graceful gloves',
+		inputItems: {
+			[itemID('Mark of grace')]: 30
+		},
+		outputItems: {
+			[itemID('Graceful gloves')]: 1
+		}
+	},
+	{
+		name: 'Graceful boots',
+		inputItems: {
+			[itemID('Mark of grace')]: 40
+		},
+		outputItems: {
+			[itemID('Graceful boots')]: 1
+		}
+	},
+	{
+		name: 'Graceful cape',
+		inputItems: {
+			[itemID('Mark of grace')]: 40
+		},
+		outputItems: {
+			[itemID('Graceful cape')]: 1
+		}
+	},
+	/**
+	 * Prospector outfit
+	 */
+	{
+		name: 'Prospector helmet',
+		outputItems: {
+			[itemID('Prospector helmet')]: 1
+		},
+		inputItems: {
+			[itemID('Golden nugget')]: 40
+		}
+	},
+	{
+		name: 'Prospector jacket',
+		outputItems: {
+			[itemID('Prospector jacket')]: 1
+		},
+		inputItems: {
+			[itemID('Golden nugget')]: 60
+		}
+	},
+	{
+		name: 'Prospector legs',
+		outputItems: {
+			[itemID('Prospector legs')]: 1
+		},
+		inputItems: {
+			[itemID('Golden nugget')]: 50
+		}
+	},
+	{
+		name: 'Prospector boots',
+		outputItems: {
+			[itemID('Prospector boots')]: 1
+		},
+		inputItems: {
+			[itemID('Golden nugget')]: 30
+		}
+	},
+	{
+		name: 'Mining gloves',
+		outputItems: {
+			[itemID('Mining gloves')]: 1
+		},
+		inputItems: {
+			[itemID('Unidentified minerals')]: 60
+		}
+	},
+	{
+		name: 'Superior mining gloves',
+		outputItems: {
+			[itemID('Superior mining gloves')]: 1
+		},
+		inputItems: {
+			[itemID('Unidentified minerals')]: 120
+		}
+	},
+	{
+		name: 'Expert mining gloves',
+		outputItems: {
+			[itemID('Expert mining gloves')]: 1
+		},
+		inputItems: {
+			[itemID('Superior mining gloves')]: 1,
+			[itemID('Mining gloves')]: 1,
+			[itemID('Unidentified minerals')]: 60
+		}
+	}
+];
+
+export default SkillingSets;
